+++ src/features/activity/main/ui/MainCategoryList.tsx
