+++ src/features/activity/main/ui/MainSectionLayout.tsx
