+++ src/features/activity/ui/HeroBanner.tsx
